+++ CHANGELOG.md
@@ -2,5 +2,7 @@
 
 ## v 0.1.1
 
-- Add drop_irw() function to drop integrated reading and writing tests
-- Add drop_failing_retests() function
+- Add `drop_irw()` function to drop integrated reading and writing tests
+- Add `drop_failing_retests()` function.
+- Export `filter_exclusions()` function
+- Fix error in `filter_exclusions()` function to properly exclude recently arrived ELs only for failing Reading retests